New asset update

See merge request vajahath/miwago-backend!288

+++ src/routes/project/delete-view-project.ts
@@ -1,5 +1,6 @@
 import { RequestHandler } from 'express';
 import { Project } from '../../models/Project';
+import { Promise as BluePromise } from 'bluebird';
 
 import '../../models/ProjectCategory';
 import '../../models/ProjectSubCategory';
@@ -11,6 +12,8 @@ import {
   RequestError,
   RequestErrorType,
 } from '../../error-handler/RequestError';
+import { Favorites } from '../../models/Favorites';
+import { ProjectRequest } from '../../models/ProjectRequest';
 
 export const getProjectById: RequestHandler = async (req, res, next) => {
   try {
@@ -33,9 +36,23 @@ export const getProjectById: RequestHandler = async (req, res, next) => {
 
 export const deleteProjectById: RequestHandler = async (req, res, next) => {
   try {
-    await Project.findByIdAndRemove({
+    const isExists = await ProjectRequest.count({
+      projectId: req.query.projectId,
+    });
+    if (isExists !== 0) {
+      return res.status(409).send({
+        success: false,
+        msg: 'Project is actively used by a client',
+      });
+    }
+    const projectPromise = Project.findByIdAndRemove({
       _id: req.query.projectId,
     }).exec();
+
+    const favoritePromise = Favorites.remove({
+      projectsId: req.query.projectId,
+    });
+    await BluePromise.all([projectPromise, favoritePromise]);
     return res.status(200).send({
       success: true,
     });
